--- Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pylint = "*"
-
-[packages]
-pyvisa-py = "*"
-scikit-rf = "*"
-pyyaml = "*"
-
-[requires]
-python_version = "3.7"
